v8: much better
code was redone again from scratch
i dont remember everything i changed but here is some
The script now uses the yt_dlp library to search for YouTube videos instead of the youtube-dl library. This change improves search results and makes the script more reliable.
Searches go through a more complex algorithm to find the best video. If the video doesn't meet expectations, it will be skipped.
The script searches for the video in the background and changes the video only when the next one is found and checked.
Caching is now implemented, making it faster and more reliable. The cache is stored in an easy-to-read format for tuning the algorithm.
An override file allows manual override of the search results and cache file to ensure the correct song is played or for a casual rickroll.
UTF-8 encoding ensures the script can handle all songs and most languages.
Language support is implemented. Currently, only English and Japanese are fully supported, but others may work with varying success.
The search algorithm is now tunable to ensure the correct song is played. A script for testing and tuning is planned.
The script detects if the track is set to loop in Spotify and will loop as well.
The script detects if the song is paused and will pause the video as well.
Fullscreen playback is now enabled.
Audio is set to 1 to ensure playback, preventing Spotify complaints.
The looping Spotify audio is now set to a shorter period to be unheard on phones where volume control isn't possible.